backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib
bcrypt==3.2.2
python-multipart
sqlalchemy
databases[sqlite]
psycopg2-binary
asyncpg
python-dotenv
pydantic[email]
httpx